refactor tests to be at single place and remove gpu kernel for value_counts which is not being used any where else
